requirements.txt
httpx==0.27.2
fastapi==0.115.12
uvicorn==0.24.0
langgraph==0.0.65
pydantic==2.5.0
openai==1.3.0
pandas
rapidfuzz
python-dotenv==1.0.0